+++ src/main/java/com/marginallyclever/robotoverlord/renderpanel/OpenGLRenderPanel.java
@@ -533,7 +533,7 @@ import java.util.*;
         gl.glClear(GL3.GL_DEPTH_BUFFER_BIT | GL3.GL_STENCIL_BUFFER_BIT);
         useShaderDefault(gl);
         // 3D renderpasses
-        if (showWorldOrigin.get()) MatrixHelper.createMesh(10).render(gl);
+        if (showWorldOrigin.get()) MatrixHelper.createMesh().render(gl);
         for(EditorTool tool : editorTools) tool.render(gl,shaderDefault);
         // 2D renderpasses
         compass3d.render(gl,viewport,shaderDefault);
