+++ src/main/java/com/marginallyclever/robotoverlord/systems/physics/ode/ODEPhysicsComponent.java
@@ -70,6 +70,6 @@ public class ODEPhysicsComponent extends Component {
 				ny.x,ny.y,ny.z,p.y,
 				nz.x,nz.y,nz.z,p.z,
 				0,0,0,1);
-		MatrixHelper.createMesh(m,10).render(gl);
+		MatrixHelper.createMesh().render(gl);
 	}
 }
